feat(provider): add shouldDisplayWelcome, injectWelcomeBanner
