+++ src/extension/content-script/index.tsx
@@ -1,4 +1,3 @@
-import './injections/Welcome' // ? Inject welcome
 import './injections/Posts' // ? Inject all posts
 import './tasks' // ? AutomatedTabTask Run tasks when invoked by background page
