+++ src/social-network-provider/facebook.com/ui-provider.ts
@@ -11,6 +11,8 @@ import { getPostContentFacebook } from './tasks/getPostContent'
 import { resolveLastRecognizedIdentityFacebook } from './UI/resolveLastRecognizedIdentity'
 import { getProfileFacebook } from './tasks/getProfile'
 import { pasteIntoBioFacebook } from './tasks/pasteIntoBio'
+import { shouldDisplayWelcomeDefault } from '../../social-network/defaults/shouldDisplayWelcome'
+import { injectWelcomeBannerFacebook } from './UI/injectWelcomeBanner'
 
 defineSocialNetworkUI({
     ...sharedProvider,
@@ -22,11 +24,13 @@ defineSocialNetworkUI({
     shouldActivate() {
         return location.hostname.endsWith('facebook.com')
     },
+    shouldDisplayWelcome: shouldDisplayWelcomeDefault('facebook.com'),
     friendsRef: new ValueRef([]),
     myIdentitiesRef: new ValueRef([]),
     lastRecognizedIdentity: new ValueRef({ identifier: PersonIdentifier.unknown }),
     resolveLastRecognizedIdentity: resolveLastRecognizedIdentityFacebook,
     injectPostBox: injectPostBoxFacebook,
+    injectWelcomeBanner: injectWelcomeBannerFacebook,
     collectPeople: collectPeopleFacebook,
     taskPasteIntoPostBox: pasteIntoPostBoxFacebook,
     taskPasteIntoBio: pasteIntoBioFacebook,
